# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for RSV Infotech PO Generator
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock-items:
    get:
      operationId: listStockItems
      tags: [stock-items]
      summary: List stock items for current company
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
            enum: [product, service]
      responses:
        "200":
          description: List of stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItem"
    post:
      operationId: createStockItem
      tags: [stock-items]
      summary: Create a stock item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
  /stock-items/{id}:
    get:
      operationId: getStockItem
      tags: [stock-items]
      summary: Get a stock item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
    put:
      operationId: updateStockItem
      tags: [stock-items]
      summary: Update a stock item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStockItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
    delete:
      operationId: deleteStockItem
      tags: [stock-items]
      summary: Delete a stock item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    StockItem:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        uom:
          type: string
        type:
          type: string
          enum: [product, service]
        unitPrice:
          type: number
        stockQty:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, companyId, code, name, uom, type, isActive]
    CreateStockItemBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        uom:
          type: string
        type:
          type: string
          enum: [product, service]
        unitPrice:
          type: number
        stockQty:
          type: number
        isActive:
          type: boolean
      required: [code, name]
    UpdateStockItemBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        uom:
          type: string
        type:
          type: string
          enum: [product, service]
        unitPrice:
          type: number
        stockQty:
          type: number
        isActive:
          type: boolean